Rename folder.
