--- test/rpc/go.mod
@@ -1,17 +1,0 @@
-module github.com/IBM/kar/test/rpc
-
-go 1.15
-
-require (
-	github.com/IBM/kar/core v1.2.0 // indirect
-	github.com/Shopify/sarama v1.26.4
-	github.com/cenkalti/backoff/v4 v4.1.1
-	github.com/gomodule/redigo v1.8.5
-	github.com/google/uuid v1.1.2
-	github.com/hashicorp/golang-lru v0.5.4
-	github.com/julienschmidt/httprouter v1.3.0
-	github.com/prometheus/client_golang v1.11.0
-	golang.org/x/net v0.0.0-20210224082022-3d97a244fca7
-	k8s.io/api v0.21.3
-	k8s.io/apimachinery v0.21.3
-)
