+++ test/rpctest/run_call_test.go
@@ -14,7 +14,7 @@
 // limitations under the License.
 //
 
-package test
+package rpctest
 
 import (
 	"testing"
